+++ cli/cli/kurtosis_config/kurtosis_config_provider.go
@@ -1,6 +1,8 @@
 package kurtosis_config
 
 import (
+	"github.com/kurtosis-tech/kurtosis/cli/cli/helpers/interactive_terminal_decider"
+	"github.com/kurtosis-tech/kurtosis/cli/cli/kurtosis_config/email_collector"
 	"github.com/kurtosis-tech/kurtosis/cli/cli/kurtosis_config/resolved_config"
 	"github.com/kurtosis-tech/stacktrace"
 	"github.com/sirupsen/logrus"
@@ -34,6 +36,11 @@ func (configProvider *KurtosisConfigProvider) GetOrInitializeConfig() (*resolved
 		kurtosisConfig, err = initConfig()
 		if err != nil {
 			return nil, stacktrace.Propagate(err, "An error occurred executing init interactive config")
+		}
+
+		// only ask for input on interactive terminal
+		if interactive_terminal_decider.IsInteractiveTerminal() {
+			email_collector.AskUserForEmailAndLogIt()
 		}
 
 		if err = configProvider.configStore.SetConfig(kurtosisConfig); err != nil {