+++ cli/cli/kurtosis_config/overrides_objects/v2/kurtosis_config_v2.go
@@ -22,4 +22,5 @@ type KurtosisConfigV2 struct {
 
 	ShouldSendMetrics *bool                               `yaml:"should-send-metrics,omitempty"`
 	KurtosisClusters  map[string]*KurtosisClusterConfigV2 `yaml:"kurtosis-clusters,omitempty"`
+	CloudConfig       *KurtosisCloudConfigV2              `yaml:"cloud-config,omitempty"`
 }