+++ container-engine-lib/lib/backend_impls/kubernetes/kubernetes_kurtosis_backend/kubernetes_kurtosis_backend_enclave_functions.go
@@ -2,11 +2,8 @@ package kubernetes_kurtosis_backend
 
 import (
 	"context"
-	"encoding/json"
 	"fmt"
-	"github.com/gammazero/workerpool"
 	"github.com/kurtosis-tech/kurtosis/container-engine-lib/lib/backend_impls/kubernetes/kubernetes_kurtosis_backend/shared_helpers"
-	"github.com/kurtosis-tech/kurtosis/container-engine-lib/lib/backend_impls/kubernetes/kubernetes_manager"
 	"github.com/kurtosis-tech/kurtosis/container-engine-lib/lib/backend_impls/kubernetes/kubernetes_resource_collectors"
 	"github.com/kurtosis-tech/kurtosis/container-engine-lib/lib/backend_impls/kubernetes/object_attributes_provider/kubernetes_annotation_key_consts"
 	"github.com/kurtosis-tech/kurtosis/container-engine-lib/lib/backend_impls/kubernetes/object_attributes_provider/label_key_consts"
@@ -19,30 +16,8 @@ import (
 	"io"
 	apiv1 "k8s.io/api/core/v1"
 	applyconfigurationsv1 "k8s.io/client-go/applyconfigurations/core/v1"
-	"os"
-	"path"
 	"strings"
 	"time"
-)
-
-const (
-	// Name to give the file that we'll write for storing specs of pods, containers, etc.
-	podSpecFilename             = "spec.json"
-	containerLogsFilenameSuffix = ".log"
-
-	// Permissions for the files & directories we create as a result of the dump
-	createdDirPerms  os.FileMode = 0755
-	createdFilePerms os.FileMode = 0644
-
-	numPodsToDumpAtOnce = 20
-
-	shouldFollowPodLogsWhenDumping        = false
-	shouldAddTimestampsWhenDumpingPodLogs = true
-
-	enclaveDumpJsonSerializationIndent = "  "
-	enclaveDumpJsonSerializationPrefix = ""
-
-	dumpPodErrorTitle = "Pod"
 )
 
 // TODO: MIGRATE THIS FOLDER TO USE STRUCTURE OF USER_SERVICE_FUNCTIONS MODULE
@@ -59,11 +34,6 @@ type enclaveKubernetesResources struct {
 	// Not technically resources that define an enclave, but we need them for
 	// StopEnclave
 	services []apiv1.Service
-}
-
-type dumpPodResult struct {
-	podName string
-	err     error
 }
 
 // ====================================================================================================
@@ -226,7 +196,7 @@ func (backend *KubernetesKurtosisBackend) StopEnclaves(
 
 			if len(errorsByPodName) > 0 {
 				combinedErrorTitle := fmt.Sprintf("Namespace %v - Pod", namespaceName)
-				combinedError := buildCombinedError(errorsByPodName, combinedErrorTitle)
+				combinedError := shared_helpers.BuildCombinedError(errorsByPodName, combinedErrorTitle)
 				erroredEnclaveIds[enclaveId] = stacktrace.Propagate(
 					combinedError,
 					"An error occurred removing one or more pods in namespace '%v' for enclave with ID '%v'",
@@ -254,7 +224,7 @@ func (backend *KubernetesKurtosisBackend) StopEnclaves(
 
 			if len(errorsByServiceName) > 0 {
 				combinedErrorTitle := fmt.Sprintf("Namespace %v - Service", namespaceName)
-				combinedError := buildCombinedError(errorsByServiceName, combinedErrorTitle)
+				combinedError := shared_helpers.BuildCombinedError(errorsByServiceName, combinedErrorTitle)
 				erroredEnclaveIds[enclaveId] = stacktrace.Propagate(
 					combinedError,
 					"An error occurred removing one or more service's selectors in namespace '%v' for enclave with ID '%v'",
@@ -281,57 +251,15 @@ func (backend *KubernetesKurtosisBackend) DumpEnclave(ctx context.Context, encla
 		return stacktrace.NewError("Cannot dump enclave '%v' because no Kubernetes namespace exists for it", enclaveUuid)
 	}
 
-	// Create output directory
-	if _, err := os.Stat(outputDirpath); !os.IsNotExist(err) {
-		return stacktrace.NewError("Cannot create output directory at '%v'; directory already exists", outputDirpath)
-	}
-	if err := os.Mkdir(outputDirpath, createdDirPerms); err != nil {
-		return stacktrace.Propagate(err, "An error occurred creating output directory at '%v'", outputDirpath)
-	}
-
 	podsToDump := kubernetesResources.pods
 	if podsToDump == nil {
 		podsToDump = []apiv1.Pod{}
 	}
 
-	workerPool := workerpool.New(numPodsToDumpAtOnce)
-	resultErrsChan := make(chan dumpPodResult, len(podsToDump))
-	for _, pod := range podsToDump {
-		/*
-			!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!! WARNING !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
-			It's VERY important that the actual `func()` job function get created inside a helper function!!
-			This is because variables declared inside for-loops are created BY REFERENCE rather than by-value, which
-				means that if we inline the `func() {....}` creation here then all the job functions would get a REFERENCE to
-				any variables they'd use.
-			This means that by the time the job functions were run in the worker pool (long after the for-loop finished)
-				then all the job functions would be using a reference from the last iteration of the for-loop.
-
-			For more info, see the "Variables declared in for loops are passed by reference" section of:
-				https://www.calhoun.io/gotchas-and-common-mistakes-with-closures-in-go/ for more details
-			!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!! WARNING !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
-		*/
-		jobToSubmit := createDumpPodJob(
-			ctx,
-			backend.kubernetesManager,
-			namespace.Name,
-			pod,
-			outputDirpath,
-			resultErrsChan,
-		)
-		workerPool.Submit(jobToSubmit)
-	}
-	workerPool.StopWait()
-	close(resultErrsChan)
-
-	resultErrorsByPodName := map[string]error{}
-	for podResult := range resultErrsChan {
-		resultErrorsByPodName[podResult.podName] = podResult.err
+	if err = shared_helpers.DumpNamespacePods(ctx, backend.kubernetesManager, namespace, podsToDump, outputDirpath); err != nil {
+		return stacktrace.Propagate(err, "An error occurred dumping pods '%+v' in namespace '%v'", podsToDump, namespace.GetName())
 	}
 
-	if len(resultErrorsByPodName) > 0 {
-		combinedErr := buildCombinedError(resultErrorsByPodName, dumpPodErrorTitle)
-		return combinedErr
-	}
 	return nil
 }