+++ core/server/api_container/server/startosis_engine/startosis_validator/validator_environment_test.go
@@ -7,15 +7,20 @@ import (
 )
 
 const (
-	testBarService = service.ServiceName("bar")
-	fooPortId      = "foo"
-	fizzPortId     = "fizz"
-	invalidPortId  = "invalid"
+	testBarService                = service.ServiceName("bar")
+	fooPortId                     = "foo"
+	fizzPortId                    = "fizz"
+	invalidPortId                 = "invalid"
+	availableMemoryInBytes        = 12000
+	availableCpuInMilliCores      = 4231
+	isResourceInformationComplete = true
+	tooMuchMemory                 = 120000
+	tooMuchCpu                    = 5000
 )
 
 func TestMultiplePortIdsForValidation(t *testing.T) {
 	emptyInitialMapping := map[service.ServiceName][]string{}
-	validatorEnvironment := NewValidatorEnvironment(false, nil, nil, emptyInitialMapping)
+	validatorEnvironment := NewValidatorEnvironment(false, nil, nil, emptyInitialMapping, availableCpuInMilliCores, availableMemoryInBytes, isResourceInformationComplete)
 	portIds := []string{
 		fooPortId,
 		fizzPortId,
@@ -27,4 +32,6 @@ func TestMultiplePortIdsForValidation(t *testing.T) {
 	validatorEnvironment.RemoveServiceFromPrivatePortIDMapping(testBarService)
 	require.False(t, validatorEnvironment.DoesPrivatePortIDExistForService(fooPortId, testBarService))
 	require.False(t, validatorEnvironment.DoesPrivatePortIDExistForService(fizzPortId, testBarService))
+	require.Error(t, validatorEnvironment.HasEnoughCPU(tooMuchCpu, testBarService))
+	require.Error(t, validatorEnvironment.HasEnoughMemory(tooMuchMemory, testBarService))
 }